docs: refine documentation for igist.el
This commit refines the documentation for the `igist.el` file. It adds
more detailed explanations of the major modes `igist-list-mode` and
`igist-explore-mode`. It also provides information about the
configuration variables related to these modes and the available menu
commands.

The commit also mentions the custom variables related to these modes
and provides an overview of the commands available in
the `igist-table-menu` and `igist-dispatch` menus.