feat(igist.el): add cloning gist functionality
Added the ability to clone a GitHub gist to a specified directory.

This new `igist-clone-gist` function asks for a gist ID,
destination directory, and clone protocol, then clones the
gist accordingly.

Also, the 'igist-clone-default-directory' custom variable introduced
to control the default cloning directory and a new keyboard
mapping "L" for this cloning command added into the package.

Add a new keyboard mapping "L" for this cloning command in igist list
mode.