refactor: remove unused igist-render-files and igist-tabulated-list--column-number
Remove the igist-render-files function that was previously used to
render files and join them with newlines and padding, as well as the
igist-tabulated-list--column-number function that determined the
column number of a given name in the igist-table-list-format. These
functions are no longer used and have been removed to clean up the
codebase.